Share provider upload session locks
